nextjs-flask/api/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
langchain-openai==0.2.9
langchain==0.3.7
supabase==2.10.0
flask-cors==3.0.10
psycopg2==2.9.10
langchain-community==0.3.7
pinecone==3.1.0
firecrawl-py==1.5.0
pypdf==5.1.0
docx2txt==0.8
langchain-text-splitters==0.3.2
Werkzeug==3.1.3
python-docx==1.1.2